tests/wpt/tests/css/css-flexbox/position-relative-percentage-top-005.html

<!DOCTYPE html>
<title>CSS Flexbox: percentage top on relatively positioned elements in non-stretch flex items</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#definite-sizes">
<link rel="help" href="https://drafts.csswg.org/css-position-3/#valdef-top-percentage">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
body {
    margin: 0;
}

.green {
    width: 20px;
    height: 20px;
    background: green;
    position: relative;
}
</style>

<!-- 1. Row flex, align-items:flex-end, child has height:100%, top: 50% -->
<div style="display: flex; height: 200px;">
    <div style="display: flex; align-items: flex-end;">
        <div id="t1-item" style="height: 100%;">
            <div class="green" id="t1" style="top: 50%;"></div>
        </div>
    </div>
</div>

<!-- 2. Row flex, align-items:flex-end, child has auto height, top: 50% -->
<div style="display: flex; height: 200px;">
    <div style="display: flex; align-items: flex-end;">
        <div id="t2-item">
            <div class="green" id="t2" style="top: 50%;"></div>
        </div>
    </div>
</div>

<!-- 3. Row flex, align-items:center, child has height:100%, top: 50% -->
<div style="display: flex; height: 200px;">
    <div style="display: flex; align-items: center;">
        <div id="t3-item" style="height: 100%;">
            <div class="green" id="t3" style="top: 50%;"></div>
        </div>
    </div>
</div>

<!-- 4. Row flex, align-items:flex-start, child has height:50%, top: 50% -->
<div style="display: flex; height: 200px;">
    <div style="display: flex; align-items: flex-start;">
        <div id="t4-item" style="height: 50%;">
            <div class="green" id="t4" style="top: 50%;"></div>
        </div>
    </div>
</div>

<!-- 5. Nested flex: outer stretched, inner flex-end, child has height:100%, top: 50% -->
<div style="display: flex; height: 200px;">
    <div>
        <div style="display: flex; align-items: flex-end; height: 100%;">
            <div id="t5-item" style="height: 100%;">
                <div class="green" id="t5" style="top: 50%;"></div>
            </div>
        </div>
    </div>
</div>

<!-- 6. Row flex, align-items:flex-end, child has fixed height:100px, top: 50% -->
<div style="display: flex; height: 200px;">
    <div style="display: flex; align-items: flex-end;">
        <div id="t6-item" style="height: 100px;">
            <div class="green" id="t6" style="top: 50%;"></div>
        </div>
    </div>
</div>

<!-- 7. Row flex, align-items:flex-end, child has height:100%, container has auto height, top: 50% -->
<div style="display: flex;">
    <div style="display: flex; align-items: flex-end;">
        <div id="t7-item" style="height: 100%;">
            <div class="green" id="t7" style="top: 50%;"></div>
        </div>
    </div>
</div>

<!-- 8. Column flex with definite height, flex item with height:50%, top: 50% -->
<div style="display: flex; flex-direction: column; height: 200px;">
    <div id="t8-item" style="height: 50%;">
        <div class="green" id="t8" style="top: 50%;"></div>
    </div>
</div>

<script>
function offsetFromItem(greenId, itemId) {
    var green = document.getElementById(greenId);
    var item = document.getElementById(itemId);
    return green.getBoundingClientRect().top - item.getBoundingClientRect().top;
}

test(function() {
    assert_equals(offsetFromItem('t1', 't1-item'), 100);
}, 'Row flex, flex-end item with height:100%, top: 50%');

test(function() {
    assert_equals(offsetFromItem('t2', 't2-item'), 0);
}, 'Row flex, flex-end item with auto height, top: 50%');

test(function() {
    assert_equals(offsetFromItem('t3', 't3-item'), 100);
}, 'Row flex, center item with height:100%, top: 50%');

test(function() {
    assert_equals(offsetFromItem('t4', 't4-item'), 50);
}, 'Row flex, flex-start item with height:50%, top: 50%');

test(function() {
    assert_equals(offsetFromItem('t5', 't5-item'), 100);
}, 'Nested flex, outer stretched, inner flex-end with height:100%, top: 50%');

test(function() {
    assert_equals(offsetFromItem('t6', 't6-item'), 50);
}, 'Row flex, flex-end item with fixed height:100px, top: 50%');

test(function() {
    assert_equals(offsetFromItem('t7', 't7-item'), 10);
}, 'Row flex, flex-end item with height:100%, auto-height container, top: 50%');

test(function() {
    assert_equals(offsetFromItem('t8', 't8-item'), 50);
}, 'Column flex with definite height, item with height:50%, top: 50%');
</script>
